reporter: Move report tabel model classes to the utils package
As they are not reporters themselves.
